add 32 bit pcm capability for I2S splitter and mixer

+++ c/hw_drivers/I2S/I2S_i94xxx/src/I2S_i94xxx.h
@@ -16,7 +16,7 @@
 
 struct I2S_i94xxx_cfg_t {
 	struct dev_desc_t *   src_clock;
-	uint8_t   	clock_mode;
+	uint32_t   	clock_mode;
 	uint32_t	sample_rate;
 	uint8_t		num_of_bytes_in_word;
 	uint8_t		DI_pin;
